feat: Move MockupTools implementation to HtmlTool
